pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Adoraxuu <adoraxu01@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
ruff = "^0.4.1"
django-taggit = "^5.0.1"
django-allauth = {extras = ["socialaccount"], version = "^0.63.1"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
